Separated the Scan to Pointcloud node to a header and c++ class
